+++ client/main.scss
@@ -1,40 +1,13 @@
-// Main stylesheet imports
-// --------------------------------------------------
-
-// Fonts
 @import '{}/imports/ui/stylesheets/fonts';
-
-// Base colors
 @import '{}/imports/ui/stylesheets/base-colors';
-
-// Bootstrap framework
 @import '{}/imports/ui/stylesheets/bootstrap';
-
-// Variables
 @import '{}/imports/ui/stylesheets/variables';
-
-// Mixins
-@import '{}/imports/ui/stylesheets/mixins';
-
-// Main layout
+@import '{}/imports/ui/stylesheets/mixins/index';
 @import '{}/imports/ui/stylesheets/layout';
-
-// Global styles
 @import '{}/imports/ui/stylesheets/global';
-
-// Noscript
 @import '{}/imports/ui/stylesheets/noscript';
-
-// UI Components
-@import '{}/imports/ui/stylesheets/components';
-
-// Right-to-left styles
+@import '{}/imports/ui/stylesheets/components/index';
 @import '{}/imports/ui/stylesheets/rtl';
-
-// Template styles
 @import '{}/imports/ui/components/index';
-
 @import '{}/imports/ui/pages/index';
-
-// Stylesheets from old structure outside of imports
 @import 'views/views';
