+++ imports/ui/components/filter/filter.js
@@ -1,8 +1,14 @@
+import { Session } from 'meteor/session';
+import { Template } from 'meteor/templating';
+import { $ } from 'meteor/jquery';
+
 import FilterPreview from '/imports/ui/lib/filter-preview.js';
 import ScssVars from '/imports/ui/lib/scss-vars.js';
 import '/imports/StringTools.js';
 
 import '/imports/ui/components/categories/categories.js';
+
+import './filter.html';
 
 Template.filter.onCreated(function() {
 	this.stateFilters =
