+++ imports/ui/components/index.scss
@@ -8,6 +8,7 @@
 @import './courses/course-members/course-members';
 @import './courses/course-roles/course-roles';
 @import './editable/editable';
+@import './filter/filter';
 @import './footer/footer';
 @import './introduction/introduction';
 @import './loading/loading';
@@ -15,6 +16,8 @@
 @import './messages/messages';
 @import './navbar/navbar';
 @import './price-policy/price-policy';
+@import './region-selection/region-selection';
+@import './region-splash/region-splash';
 @import './region-tag/region-tag';
 @import './report/report';
 @import './sharing/sharing';
