+++ imports/ui/components/region-splash/region-splash.scss
