+++ imports/ui/layouts/app-body.js
@@ -11,6 +11,7 @@ import '/imports/ui/components/footer/footer.js';
 import '/imports/ui/components/introduction/introduction.js';
 import '/imports/ui/components/messages/messages.js';
 import '/imports/ui/components/navbar/navbar.js';
+import '/imports/ui/components/region-splash/region-splash.js';
 import '/imports/ui/components/translate-info/translate-info.js';
 
 import './app-body.html';
