+++ imports/ui/pages/index.scss
@@ -1,6 +1,7 @@
 @import './calendar/calendar';
 @import './course-details/course-details';
 @import './faq/faq';
+@import './find/find';
 @import './frames/events-frame/events-frame';
 @import './frames/schedule-frame/schedule-frame';
 @import './frames/week-frame/week-frame';
