--- imports/ui/stylesheets/components.scss
@@ -1,16 +1,0 @@
-@import 'components/actions';
-@import 'components/buttons';
-@import 'components/checkbox';
-@import 'components/container-sm';
-@import 'components/detail-pages';
-@import 'components/dropdown';
-@import 'components/dialog-dropdown';
-@import 'components/edit-pages';
-@import 'components/event-caption';
-@import 'components/form-groups';
-@import 'components/frame-list';
-@import 'components/list-group';
-@import 'components/modals';
-@import 'components/nav-tabs';
-@import 'components/tag';
-@import 'components/tooltip';
