--- imports/ui/stylesheets/mixins.scss
@@ -1,6 +1,0 @@
-@import 'mixins/buttons';
-@import 'mixins/container';
-@import 'mixins/ellipsis';
-@import 'mixins/list-unstyled';
-@import 'mixins/rtl';
-@import 'mixins/tag';
