+++ server/extracts.msgfmt~
@@ -1,1 +1,1 @@
-[{"roles.message.placeholder":{"key":"roles.message.placeholder","text":"My interests...","file":"imports/ui/components/courses/course-members/course-members.js","line":71,"mtime":1509497472670,"template":"unknown","ctime":1509497472670}},{"extractedAt":1509497483197,"updatedAt":1509497472670}]
+[{"filterCaptions.is-proposal":{"key":"filterCaptions.is-proposal","text":"Proposal","file":"imports/ui/components/filter/filter.js","line":18,"mtime":1509498854496,"template":"unknown","ctime":1509498854496},"filterCaptions.resting.label":{"key":"filterCaptions.resting.label","text":"Resting","file":"imports/ui/components/filter/filter.js","line":30,"mtime":1509498854496,"template":"unknown","ctime":1509498854496},"filterCaptions.resting.title":{"key":"filterCaptions.resting.title","text":"Courses with passed but without upcoming events","file":"imports/ui/components/filter/filter.js","line":31,"mtime":1509498854496,"template":"unknown","ctime":1509498854496},"filterCaptions.showProposal":{"key":"filterCaptions.showProposal","text":"Show all proposed courses","file":"imports/ui/components/filter/filter.js","line":19,"mtime":1509498854496,"template":"unknown","ctime":1509498854496},"filterCaptions.upcoming.label":{"key":"filterCaptions.upcoming.label","text":"Upcoming","file":"imports/ui/components/filter/filter.js","line":24,"mtime":1509498854496,"template":"unknown","ctime":1509498854496},"filterCaptions.upcoming.title":{"key":"filterCaptions.upcoming.title","text":"Show all courses with upcoming events","file":"imports/ui/components/filter/filter.js","line":25,"mtime":1509498854496,"template":"unknown","ctime":1509498854496},"find.needsHost":{"key":"find.needsHost","text":"Looking for a host","file":"imports/ui/components/filter/filter.js","line":136,"mtime":1509498854496,"template":"unknown","ctime":1509498854496},"find.needsMentor":{"key":"find.needsMentor","text":"Looking for a mentor","file":"imports/ui/components/filter/filter.js","line":132,"mtime":1509498854496,"template":"unknown","ctime":1509498854496},"find.needsOrganizer":{"key":"find.needsOrganizer","text":"Looking for an organizer","file":"imports/ui/components/filter/filter.js","line":128,"mtime":1509498854496,"template":"unknown","ctime":1509498854496}},{"extractedAt":1509498866404,"updatedAt":1509498854496}]
